ensure components always go into a folder
When selecting a custom folder to install into, if you don't
explicityly create and rename a new folder, everything gets
installed into what would logically be the parent folder.
This forces a new folder to be created and renamed to the
project name, but this means that the default install is installed
as a subfolder with the same name.